feat(process/scanner): offload scanning to a worker thread
